requirements.txt
pytest
locust
flask
mysql-connector-python